Add/remove block framework
Blocks can now be added/removed while playing. Left click to add blocks
and right click to remove them. This framework will be updated as
features are added.

Also, new bug discovered. Occasionally, the Timemap._map variable will
be treated as if it does not exist. I do not know why yet. It appears to
be related to saving scripts, as this bug does not trigger on built
versions.